+++ projects/ngx-opendesign/src/public-api.ts
@@ -63,6 +63,7 @@ export * from './lib/reorder-list/reorder-list';
 export * from './lib/drawer/drawer';
 export * from './lib/media-player/media-player';
 export * from './lib/slider/slider';
+export * from './lib/step-slider/step-slider';
 export * from './lib/time-picker/time-picker';
 export * from './lib/date-range-picker/date-range-picker';
 export * from './lib/color-picker/color-picker';